+++ include/solvers/fluid_dynamics_matrix_free_operators.h
@@ -382,6 +382,24 @@ public:
     const VectorType &time_derivative_previous_solutions);
 
   /**
+   * @brief Store the values of the vector containing the ALE velocity to use
+   * them in the Jacobian and residual cell integrals. This function is called
+   * when the mortar feature is enabled.
+   *
+   * @param[in] mapping Describes the transformations from unit to real cell.
+   * @param[in] radius Radius at the mortar interface in the direction
+   * perpendicular to the rotation axis.
+   * @param[in] center_of_rotation Center of rotation of the rotor domain.
+   * @param[in] rotor_angular_velocity  Angular velocity of the rotor domain.
+   */
+  void
+  evaluate_velocity_ale(
+    const Mapping<dim>                             &mapping,
+    const double                                    radius,
+    const Point<dim>                                center_of_rotation,
+    std::shared_ptr<Functions::ParsedFunction<dim>> rotor_angular_velocity);
+
+  /**
    * @brief Store the values of the source term calculated if dynamic control
    * is enabled in the appropriate structure.
    *
@@ -743,6 +761,13 @@ protected:
    * @brief Flag to turn the computation of mortar coupling terms on or off.
    */
   bool enable_mortar;
+
+  /**
+   * @brief Table with correct alignment for vectorization to store the values
+   * of the ALE velocity used in mortar coupling terms.
+   *
+   */
+  Table<2, Tensor<1, dim, VectorizedArray<number>>> velocity_ale;
 
   /**
    * @brief Vector with the constrained indices used for the local smoothing approach.