+++ source/solvers/fluid_dynamics_matrix_free_operators.cc
@@ -1080,6 +1080,72 @@ NavierStokesOperatorBase<dim, number>::
 
 template <int dim, typename number>
 void
+NavierStokesOperatorBase<dim, number>::evaluate_velocity_ale(
+  const Mapping<dim>                             &mapping,
+  const double                                    radius,
+  const Point<dim>                                center_of_rotation,
+  std::shared_ptr<Functions::ParsedFunction<dim>> rotor_angular_velocity)
+{
+  this->timer.enter_subsection("operator::evaluate_velocity_ale");
+
+  const unsigned int n_cells =
+    matrix_free.n_cell_batches() + matrix_free.n_ghost_cell_batches();
+
+  FECellIntegrator integrator(matrix_free);
+  matrix_free.update_mapping(mapping);
+
+  // 1. Precompute values on cells:
+
+  // Set appropriate size for tables
+  velocity_ale.reinit(n_cells, integrator.n_q_points);
+
+  // Get updated rotor angular velocity
+  rotor_angular_velocity->set_time(
+    this->simulation_control->get_current_time());
+  const double rotor_angular_velocity_value =
+    rotor_angular_velocity->value(Point<dim>());
+
+  for (unsigned int cell = 0; cell < n_cells; ++cell)
+    {
+      integrator.reinit(cell);
+      for (auto lane = 0u;
+           lane < matrix_free.n_active_entries_per_cell_batch(cell);
+           lane++)
+        {
+          // Compute radial distance from the current cell
+          const auto cell_center =
+            matrix_free.get_cell_iterator(cell, lane)->center();
+          const auto radius_current = cell_center.distance(center_of_rotation);
+
+          // Use prescribed rotor angular velocity only if cell is part of the
+          // rotor
+          double cell_rotor_angular_velocity;
+          if (radius_current > radius)
+            cell_rotor_angular_velocity = 0.0;
+          else
+            cell_rotor_angular_velocity = rotor_angular_velocity_value;
+
+          // Compute linear velocity at quadrature points
+          for (const auto q : integrator.quadrature_point_indices())
+            {
+              // Assumption in 3D case: rotation axis is in z
+              // TODO generalize rotation axis
+              const auto x = integrator.quadrature_point(q)[0][lane];
+              const auto y = integrator.quadrature_point(q)[1][lane];
+
+              this->velocity_ale[cell][q][0][lane] =
+                -cell_rotor_angular_velocity * y;
+              this->velocity_ale[cell][q][1][lane] =
+                cell_rotor_angular_velocity * x;
+            }
+        }
+    }
+
+  this->timer.leave_subsection("operator::evaluate_velocity_ale");
+}
+
+template <int dim, typename number>
+void
 NavierStokesOperatorBase<dim, number>::update_beta_force(
   const Tensor<1, dim> &beta_force)
 {
@@ -1402,6 +1468,10 @@ NavierStokesStabilizedOperator<dim, number>::do_cell_integral_local(
   bool transient =
     (is_bdf(this->simulation_control->get_assembly_method())) ? true : false;
 
+  // To identify whether mortar feature is enabled (use local variable for
+  // efficiency)
+  const bool enable_mortar = this->enable_mortar;
+
   const double kinematic_viscosity =
     this->properties_manager->get_rheology()->get_kinematic_viscosity();
 
@@ -1449,6 +1519,10 @@ NavierStokesStabilizedOperator<dim, number>::do_cell_integral_local(
         previous_time_derivatives =
           this->time_derivatives_previous_solutions(cell, q);
 
+      Tensor<1, dim, VectorizedArray<number>> u_ale;
+      if (enable_mortar)
+        u_ale = this->velocity_ale[cell][q];
+
       // Get stabilization parameter
       const auto tau      = this->stabilization_parameter[cell][q];
       const auto tau_lsic = this->stabilization_parameter_lsic[cell][q];
@@ -1468,6 +1542,10 @@ NavierStokesStabilizedOperator<dim, number>::do_cell_integral_local(
               // +(v,(u·∇)δu + (δu·∇)u)
               value_result[i] += gradient[i][k] * previous_values[k] +
                                  previous_gradient[i][k] * value[k];
+
+              // -(v, (u_ale·∇)δu)
+              if (enable_mortar)
+                value_result[i] -= gradient[i][k] * u_ale[k];
             }
           // +(v,∂t δu)
           if (transient)
@@ -1484,6 +1562,10 @@ NavierStokesStabilizedOperator<dim, number>::do_cell_integral_local(
                 tau * (-kinematic_viscosity * hessian_diagonal[i][k] +
                        gradient[i][k] * previous_values[k] +
                        previous_gradient[i][k] * value[k]);
+
+              // -(u_ale·∇δu)·τ∇q
+              if (enable_mortar)
+                gradient_result[dim][i] -= tau * (gradient[i][k] * u_ale[k]);
             }
           // +(∂t δu)·τ∇q
           if (transient)
@@ -1506,6 +1588,22 @@ NavierStokesStabilizedOperator<dim, number>::do_cell_integral_local(
                     (gradient[i][l] * previous_values[l] +
                      previous_gradient[i][l] * value[l] -
                      kinematic_viscosity * hessian_diagonal[i][l]);
+
+                  if (enable_mortar)
+                    {
+                      // -((u_ale·∇)δu)τ(u·∇)v
+                      gradient_result[i][k] -=
+                        tau * previous_values[k] * gradient[i][l] * u_ale[l];
+                      // (-(u_ale·∇)δu)τ(-u_ale·∇)v
+                      gradient_result[i][k] +=
+                        tau * u_ale[k] * gradient[i][l] * u_ale[l];
+                      // -((u·∇)δu + (δu·∇)u - ν∆δu)τ(u_ale·∇)v
+                      gradient_result[i][k] -=
+                        tau * u_ale[k] *
+                        (gradient[i][l] * previous_values[l] +
+                         previous_gradient[i][l] * value[l] -
+                         kinematic_viscosity * hessian_diagonal[i][l]);
+                    }
                 }
               // +(∇δp)τ(u·∇)v
               gradient_result[i][k] +=
@@ -1516,6 +1614,15 @@ NavierStokesStabilizedOperator<dim, number>::do_cell_integral_local(
                 gradient_result[i][k] +=
                   tau * previous_values[k] * ((*bdf_coefs)[0] * value[i]);
 
+              if (enable_mortar)
+                {
+                  // -(∇δp)τ(u_ale·∇)v
+                  gradient_result[i][k] -= tau * u_ale[k] * (gradient[dim][i]);
+                  // -(∂t δu)τ(u_ale·∇)v
+                  if (transient)
+                    gradient_result[i][k] -=
+                      tau * u_ale[k] * ((*bdf_coefs)[0] * value[i]);
+                }
 
               // Part 2
               for (unsigned int l = 0; l < dim; ++l)
@@ -1525,6 +1632,11 @@ NavierStokesStabilizedOperator<dim, number>::do_cell_integral_local(
                     tau * value[k] *
                     (previous_gradient[i][l] * previous_values[l] -
                      kinematic_viscosity * previous_hessian_diagonal[i][l]);
+
+                  // -((u_ale·∇)u)τ(δu·∇)v
+                  if (enable_mortar)
+                    gradient_result[i][k] -=
+                      tau * value[k] * previous_gradient[i][l] * u_ale[l];
                 }
               // +(∇p - f)τ(δu·∇)v
               gradient_result[i][k] +=
@@ -1630,6 +1742,10 @@ NavierStokesStabilizedOperator<dim, number>::local_evaluate_residual(
         (is_bdf(this->simulation_control->get_assembly_method())) ? true :
                                                                     false;
 
+      // To identify whether mortar feature is enabled (use local variable for
+      // efficiency)
+      const bool enable_mortar = this->enable_mortar;
+
       // Vector for BDF coefficients
       const Vector<double> *bdf_coefs;
       if (transient)
@@ -1664,6 +1780,10 @@ NavierStokesStabilizedOperator<dim, number>::local_evaluate_residual(
             previous_time_derivatives =
               this->time_derivatives_previous_solutions(cell, q);
 
+          Tensor<1, dim, VectorizedArray<number>> u_ale;
+          if (enable_mortar)
+            u_ale = this->velocity_ale[cell][q];
+
           // Get stabilization parameter
           const auto tau      = this->stabilization_parameter[cell][q];
           const auto tau_lsic = this->stabilization_parameter_lsic[cell][q];
@@ -1696,6 +1816,10 @@ NavierStokesStabilizedOperator<dim, number>::local_evaluate_residual(
                 {
                   // +(v,(u·∇)u)
                   value_result[i] += gradient[i][k] * value[k];
+
+                  // -(v, (u_ale·∇)u)
+                  if (enable_mortar)
+                    value_result[i] -= gradient[i][k] * u_ale[k];
                 }
             }
 
@@ -1708,6 +1832,10 @@ NavierStokesStabilizedOperator<dim, number>::local_evaluate_residual(
                   gradient_result[dim][i] +=
                     tau * (-kinematic_viscosity * hessian_diagonal[i][k] +
                            gradient[i][k] * value[k]);
+
+                  // -((u_ale·∇)u)·τ∇q
+                  if (enable_mortar)
+                    gradient_result[dim][i] -= tau * gradient[i][k] * u_ale[k];
                 }
               // +(-f)·τ∇q
               gradient_result[dim][i] += tau * (-source_value[i]);
@@ -1735,6 +1863,26 @@ NavierStokesStabilizedOperator<dim, number>::local_evaluate_residual(
                       // + ((u·∇)u)τ(u·∇)v
                       gradient_result[i][k] +=
                         tau * value[k] * gradient[i][l] * value[l];
+
+                      if (enable_mortar)
+                        {
+                          // -((u_ale·∇)u)τ(u·∇)v
+                          gradient_result[i][k] -=
+                            tau * value[k] * gradient[i][l] * u_ale[l];
+
+                          // -(-ν∆u)τ(u_ale·∇)v
+                          gradient_result[i][k] += tau * kinematic_viscosity *
+                                                   u_ale[k] *
+                                                   hessian_diagonal[i][l];
+
+                          // -((u·∇)u)τ(u_ale·∇)v
+                          gradient_result[i][k] -=
+                            tau * u_ale[k] * gradient[i][l] * value[l];
+
+                          // -(-(u_ale·∇)u)τ(u_ale·∇)v
+                          gradient_result[i][k] +=
+                            tau * u_ale[k] * gradient[i][l] * u_ale[l];
+                        }
                     }
                   // + (∇p - f)τ(u·∇)v
                   gradient_result[i][k] +=
@@ -1745,6 +1893,19 @@ NavierStokesStabilizedOperator<dim, number>::local_evaluate_residual(
                     gradient_result[i][k] += tau * value[k] *
                                              ((*bdf_coefs)[0] * value[i] +
                                               previous_time_derivatives[i]);
+
+                  if (enable_mortar)
+                    {
+                      // - (∇p - f)τ(u_ale·∇)v
+                      gradient_result[i][k] -=
+                        tau * u_ale[k] * (gradient[dim][i] - source_value[i]);
+
+                      // - (∂t u)τ(u_ale·∇)v
+                      if (transient)
+                        gradient_result[i][k] -= tau * u_ale[k] *
+                                                 ((*bdf_coefs)[0] * value[i] +
+                                                  previous_time_derivatives[i]);
+                    }
                 }
             }